refactor: cardList 컴포넌트의 mockDatas 를 data/listPage.js 파일로 분리
